Refactor record serialization logic into a Record::Readable role
The Record resource is now more generally usable, such as for resources
backed by records but which don't want to use a serialized
representation.

The name "Readable" may not be quite right, but it matches the existing
naming of similar roles (i.e. Writable).  For now, it's fine.  In the
future, it may want to become more specific.