Use Web::Simple as dispatcher
When processing comes to Resources layer, PATH_INFO does not contain just
the /:id anymore -- due the use of Web::Simple and not simply mount() call.

The REST path is just a prefix, so it is not necessary that we write every
single route rule containing such prefix. What matters for REST API core is
the stuff after the base path, that is, stuff like /tickets/1. Hence the use
of ReverseProxyPath.