Replace PUT with PATCH
PUT is meant to specify a complete replacement in the content; PATCH is
for specifying partial modifications. Since RT's Update() routines take
partial modifications and not complete replacements, PATCH is the more
correct method.

https://tools.ietf.org/html/rfc5789

"The difference between the PUT and PATCH requests is reflected in the
way the server processes the enclosed entity to modify the resource
identified by the Request-URI. In a PUT request, the enclosed entity is
considered to be a modified version of the resource stored on the origin
server, and the client is requesting that the stored version be
replaced. With PATCH, however, the enclosed entity contains a set of
instructions describing how a resource currently residing on the origin
server should be modified to produce a new version."