Allow web session and token-based auth
If you have a valid session cookie, we now use it. This makes it
possible to (among other things) use REST2 from JS, instead of having to
write one-off /Helpers/ endpoints.

There's also a new RT::Authen::Token extension which REST2 can use as
well. Since this plugin is likely destined for core, it doesn't check
@Plugins but instead checks to see if the feature is there.

If it looks like you're using a browser, we shuttle you over to / to log
you in rather than provide a 401 response.

Finally, the test demanding that unauthorized requests send a
WWW-Authenticate be present is removed; we don't want to prompt a
browser for basic auth. The 401 status code and response body should
suffice for non-browser clients. A sampling of other APIs shows
WWW-Authenticate isn't widely used.