ETag and If-Match support for conflict detection
Every RT::Record response now includes an ETag which can be later used
(by providing an If-Match header) to enforce "only do this update if the
server-side object hasn't been changed since I last fetched it" as well
as caching (only send me the full response if it's been changed)

https://en.wikipedia.org/wiki/HTTP_ETag

For record types that support transactions (e.g. tickets, assets, users,
groups), we use the latest transaction ID as the ETag. For other record
types (e.g. saved searches, when we add support for them) we use the
LastModified date (which isn't as robust, but is better than nothing)