Avoid generating spurious deprecation warnings in serialize_record
These deprecation warnings are meant to indicate to a developer that
their customization/extension uses a deprecated feature. That's not the
case here; we're using an abstraction.

It'd be great if we could somehow flag to RT->Deprecated that we don't
want to see such warnings but in lieu of that, suppress any deprecation
warning from calling these methods.

Added for RT::Group->Type on 4.2