Connect to the database at request start, not app init
rt-server takes pains to close down the database connection before
handing off control to the PSGI server, such that the database
connection is not preserved across the `fork`.  DB connections which
are so shared result in undefined behavior -- postgres, for instance,
will reuse the same statement ids on the handle in different
processes, resulting in errors of the form:

    DBD::Pg::st execute failed: ERROR:  prepared statement "dbdpg_p4068_1" already exists

Unfortunately, `to_app` is called _after_ that preventative
disconnection has happened, so opening the database connection there
_does_ cause it to be shared -- with the aforementioned disasterous
results.

_Some_ call to `ConnectToDatabase` is required, because (at the time
of this commit, in <= 4.4) the only other explicit call to it is done
inside the Mason handler.  Thus, without an explicit connect in this
extension, REST requests which happen to have not served a Mason page
will fail, due to lack of a database connection.

Move database connections to inside the PSGI handlers themselves, such
that the connection is opened inside the forked child, not the parent.
Just the addition in Auth is technically sufficient, since it
currently always runs before the main dispatcher -- however, we
include the connection there as a preventative measure.  This is safe
because DBIx::SearchBuilder::Handle's Connect (and thus
RT::ConnectToDatabase) is a no-op if the connection is already live.