Update REST2 docs with examples and clearer notation
Add examples of REST2.0 query with GET and POST methods, as well removing
confusing lines about GET method. Most of the endpoints have a line of
documentation about using GET 'REST/2.0/endpoint?query=<JSON>',
this is highly confusing. The GET method cannot send JSON as an
arguement and the only endpoint that accepts '?query=' is Tickets. Removing
this information and adding examples of querying of the GET method will
alleviate in the confusion.