Support to create ticket with content
Previously on ticket create, we can set ticket's fields like subject,
people, dates, custom fields, etc. But sadly we can't add content, this
commit is to fix this, so we can really do:

    curl -X POST -H "Content-Type: application/json" -u 'root:password'
        -d '{ "Queue": "General", "Subject": "Create ticket test",
            "From": "user1@example.com", "To": "rt@example.com",
            "Content": "Testing a create",
            "CustomFields": {"Severity": "Low"}}'
        'https://myrt.com/REST/2.0/ticket'

To add content on ticket create, we need to build a MIME object to feed
RT::Ticket::Create, and the following fields will be used to build the
MIME object: Subject, From, To, Cc, Date, ContentType and Content, which
HTML::Mason::Commands::MakeMIMEEntity supports now.