Show plain column value if the corresponding ...Obj is not an RT::Record
We use UID method to expand RT::Record objects. But not all the ...Obj
methods return RT::Record objects. Take LifecycleObj for example,
previously it returned an empty hash, which is not useful. Instead, we
can return the plain column value(like "default" Lifecycle) for cases
like this.

Code is also re-arranged to make the logic a bit more clear.