Improve the user experience of pagination.
Collection results which require pagination now include "pages",
"next_page" and "prev_page". The later two only if there are next
or previous pages. For example:

https://rt.example.com/REST/2.0/queues/all?per_page=5\&page=3

{
   "count" : 5,
   "pages" : 8,
   "page" : 3,
   "per_page" : 5,
   "prev_page" : "https://rt.example.com/REST/2.0/queues/all?per_page=5&page=2",
   "next_page" : "https://rt.example.com/REST/2.0/queues/all?per_page=5&page=4",
   "total" : 37,
   "items" : [ ... ]
}

Don't limit the page param by a guess of the number of pages.

Because we haven't applied the limits to the query yet, the call
to CountAll returns the total number of results. Therefore, using
that to limit the number of available of pages is useless.