Create CleanupRequest middleware to really cleanup requests for REST2
Previously we added the cleanup functionality to to_psgi_app(initially
created as an alias of "to_app") in 059ab1c61b, which never got called,
as the running order is like:

    RT::Interface::Web::Handler->PSGIWrap
    RT::Extension::REST2->PSGIWrap
    RT::Extension::REST2->to_app
    RT::Extension::REST2::Dispatcher->to_psgi_app

The call of response_cb was not quite right either as it doesn't expect
a Plack app as the first argument. The postgres issue 059ab1c61b tried
to fix was actually fixed in c3610c1a6a.

This commit wraps the functionality into a middleware and calls the more
complete version(RT::Interface::Web::Handler::CleanupRequest).